# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Travel Agency Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List umrah packages
      parameters:
        - name: duration
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get package detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
...
  /tickets/{bookingId}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get e-ticket for booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: E-ticket data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
components:
  schemas:
...
    Package:
      type: object
      required: [id, name, duration, type, price, hotel, airline, departureDate, returnDate, seatsAvailable, description, inclusions]
      properties:
        id:
          type: integer
        name:
          type: string
        duration:
          type: integer
        type:
          type: string
        price:
          type: number
        hotel:
          type: string
        hotelRating:
          type: ["integer", "null"]
        airline:
          type: string
        departureDate:
          type: string
        returnDate:
          type: string
        seatsAvailable:
          type: integer
        description:
          type: string
        inclusions:
          type: array
          items:
            type: string
        makkahNights:
          type: ["integer", "null"]
        madinahNights:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
...
    Ticket:
      type: object
      required: [ticketNumber, bookingRef, passengerName, flightNumber, airline, origin, destination, departureDate, departureTime, arrivalTime, class, seat, pnr, status, issuedAt]
      properties:
        ticketNumber:
          type: string
        bookingRef:
          type: string
        passengerName:
          type: string
        passportNumber:
          type: string
        flightNumber:
          type: string
        airline:
          type: string
        airlineCode:
          type: string
        origin:
          type: string
        originCode:
          type: string
        destination:
          type: string
        destinationCode:
          type: string
        departureDate:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: string
        class:
          type: string
        seat:
          type: string
        pnr:
          type: string
        baggage:
          type: ["string", "null"]
        status:
          type: string
        issuedAt:
          type: string
        barcode:
          type: ["string", "null"]
...